At least the display part of #3 is kinda working

+++ src/Levels.cpp
@@ -70,7 +70,7 @@ namespace LevelLogic{
 		entireData.append(10000, '0');
 		for (auto& level : levelData){
 			uint16_t charToChange = level.Y * level.X;
-			uint8_t substrID = 0;
+			uint16_t substrID = 0;
 			for (uint8_t i = 0; i < 20; i++){
 				for(uint8_t k = 0; k < 20; k++){
 					entireData[charToChange + k] = level.data[substrID];
